z algorithm implemented. started adding unit tests. project structure is a little better, but still needs improvement (probably some folders may be dropped)

+++ CodeBlocks project/algorithms/hashing/bernstein.cpp
@@ -7,7 +7,7 @@ bernstein_hasher::bernstein_hasher(int base) {
 }
 
 void bernstein_hasher::compute_powers() {
-    int power = base;
+    ULL power = base;
     ULL top = ULLONG_MAX / base;
     powers.push_back(1);
     while(1) {
@@ -37,7 +37,7 @@ ULL bernstein_hasher::next_bernstein_hash(const char * str, int index, ULL curre
 }
 
 ULL bernstein_hasher::remove_prefix_from_hash(const char * str, int index, int length, ULL current_hash) {
-    if (length > powers.size()) return NO_HASH;
+    if (length > (int)powers.size()) return NO_HASH;
     if (length < 1) return NO_HASH;
     return (current_hash - str[index]*powers[length-1]);
 }
